feat: dao 인터페이스 구현
