+++ src/main/java/wooteco/subway/dao/JdbcStationDao.java
@@ -1,0 +1,55 @@
+package wooteco.subway.dao;
+
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
+import org.springframework.stereotype.Repository;
+import wooteco.subway.domain.Station;
+
+import java.sql.PreparedStatement;
+import java.util.List;
+
+@Repository
+public class JdbcStationDao implements StationDao{
+
+    private static final RowMapper<Station> STATION_ROW_MAPPER = (rs, rowNum) -> new Station(
+            rs.getLong("id"),
+            rs.getString("name")
+    );
+    private final JdbcTemplate jdbcTemplate;
+
+    public JdbcStationDao(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    @Override
+    public Station save(Station station) {
+        String sql = "INSERT INTO station (name) VALUES(?)";
+        KeyHolder keyHolder = new GeneratedKeyHolder();
+        jdbcTemplate.update(connection -> {
+            PreparedStatement statement = connection.prepareStatement(sql, new String[]{"id"});
+            statement.setString(1, station.getName());
+            return statement;
+        }, keyHolder);
+        return new Station(keyHolder.getKey().longValue(), station.getName());
+    }
+
+    @Override
+    public List<Station> findAll() {
+        String sql = "SELECT * FROM station";
+        return jdbcTemplate.query(sql, STATION_ROW_MAPPER);
+    }
+
+    @Override
+    public Station findByName(String name) {
+        String sql = "SELECT * FROM station WHERE name=?";
+        return jdbcTemplate.queryForObject(sql, STATION_ROW_MAPPER, name);
+    }
+
+    @Override
+    public void deleteById(Long id) {
+        String sql = "DELETE FROM station WHERE id=?";
+        jdbcTemplate.update(sql, id);
+    }
+}
